Stabilize runtime_profile as one object contract
Coordinate PR #538's stage-routing table with PR #505's backend-native profile selection by reserving runtime_profile.backend_profile and keeping default/stages for stage routing. Move the stage vocabulary to an import-light module so config validation can share it without importing the full orchestrator package graph.

Constraint: PR #538 must coexist with PR #505's Codex backend profile contract while preserving stage routing.

Rejected: Keeping runtime_profile as stage-routing-only object | conflicts with PR #505's string contract.

Rejected: Duplicating stage keys inside config models | drifts from the canonical stage enum.

Confidence: high

Scope-risk: moderate

Directive: Keep backend-native profile selection under runtime_profile.backend_profile and stage routing under runtime_profile.default/stages; import stage vocabulary from ouroboros.orchestrator_stage for config validation.

Tested: uv run ruff format --check src/ouroboros/config/models.py src/ouroboros/orchestrator/stage.py src/ouroboros/orchestrator_stage.py tests/unit/orchestrator/test_stage_resolution.py

Tested: uv run ruff check src/ouroboros/config/models.py src/ouroboros/config/__init__.py src/ouroboros/orchestrator/stage.py src/ouroboros/orchestrator_stage.py tests/unit/orchestrator/test_stage_resolution.py tests/unit/config/

Tested: uv run mypy src/ouroboros/config/models.py src/ouroboros/orchestrator/stage.py src/ouroboros/orchestrator_stage.py

Tested: uv run pytest tests/unit/orchestrator/test_stage_resolution.py tests/unit/config/ -q

Not-tested: full repository test suite locally
